gui/requirements.txt
PyQt5
requests
pillow